gh-120321: Use atomics for PyGenObject.gi_frame_state
In the free threading build, use atomics to change the generator
s state. Generators can only be executed by one thread at a time and
it's an error to try to iterate over or send a value into a currently
executing generator.

The `gen_send_ex2` function needs to determine if the
`_PyEval_EvalFrame()` call resulted in the generator suspended or
returned. We can't rely on `gi_frame_state` because if the generator
suspended, a different thread could have (in theory) immediately started
executing it. Instead, route the resulting frame state through a new
`gi_out_frame_state` pointer.