gh-128421: Fix performance regeression from adding locks to frames
The recent PR gh-131479 added locking to `take_ownership` in the free
threading build. The cost is not really the locking -- that path isn't
taken frequently -- but the inlined code causes extra register spills
and slows down RETURN_VALUE, even when it's not taken.

Mark `take_ownership` as `Py_NO_INLINE` to avoid the regression.

Also limit locking in PyFrameObject to Python functions, not the C API.
PyFrame_GetCode is called frequently by coverage and tracing tools.